requirements.txt
aiogram==3.20.0.post0
python-dotenv==1.1.1
hashids==1.3.1
asyncpg==0.30.0
cachetools==6.1.0